[fix]Schema 공용으로 분리
